feat(storage): Introduce base abstract class for storage client (#1655)
This PR introduces an abstract class to support the upcoming async
client for the Python SDK. This refactor defines the public interface
and enables code sharing between sync and async client, without
introducing new logic.
